+++ examples/import_dxf_via_browser.py
@@ -1,10 +1,7 @@
 # Copyright (c) 2020, Joseph Flack
 # License: MIT License
 from typing import TYPE_CHECKING
-import io
-import base64
 import ezdxf
-from ezdxf.filemanagement import dxf_stream_info
 
 if TYPE_CHECKING:
     from ezdxf.eztypes import Drawing
@@ -23,41 +20,16 @@ def get_dxf_doc_from_upload_data(data: bytes) -> 'Drawing':
 
     """
     # Remove the mime-type and encoding info from data
-    # example: data:application/vnd.ms-excel;base64,ICAwU0VDVElPTiAgMkhFQURFUiAgOSRBQ0FEVkVSICAxQUMxMDI0ICA5JEFDQ...
+    # example: data:application/octet-stream;base64,OTk5DQpkeGZydyAwLjYuMw0KICAwDQpTRUNUSU9ODQogIDINCkhFQURFUg0KICA...
     _, data = data.split(b',')
-
-    # Decode base64 encoded data into binary data
-    binary_data = base64.b64decode(data)
-
-    # Replace windows line ending '\r\n' by universal line ending '\n'
-    binary_data = binary_data.replace(b'\r\n', b'\n')
-
-    # Read DXF file info from data, basic DXF information in the HEADER section is ASCII encoded
-    # so encoding setting here is not important for this task:
-    text = binary_data.decode('utf-8', errors='ignore')
-    stream = io.StringIO(text)
-    info = dxf_stream_info(stream)
-    stream.close()
-
-    # Use encoding info to create correct decoded text input stream for ezdxf
-    text = binary_data.decode(info.encoding, errors='ignore')
-    stream = io.StringIO(text)
-
-    # Load DXF document from data stream
-    doc = ezdxf.read(stream)
-    stream.close()
-    return doc
+    return ezdxf.decode_base64(data)
 
 
-def example_data(doc: 'Drawing') -> bytes:
-    stream = io.StringIO()
-    doc.write(stream)
-    # create binary data with windows line ending
-    binary_data = stream.getvalue().encode('utf-8').replace(b'\n', b'\r\n')
-    return b'data:application/any;base64,' + base64.encodebytes(binary_data)
+def encode_base64(doc: 'Drawing') -> bytes:
+    return b'data:application/octet-stream;base64,' + doc.encode_base64()
 
 
 if __name__ == '__main__':
-    data = example_data(ezdxf.new())
+    data = encode_base64(ezdxf.new())
     doc = get_dxf_doc_from_upload_data(data)
     print(doc.acad_release)